Update InvoiceRequest, InvoiceCheckData, and InvoiceCallbackData structures to use float64 for Amount fields, improving precision in invoice calculations.

+++ structs.go
@@ -37,13 +37,13 @@ type (
 // invoice structs
 type (
 	InvoiceRequest struct {
-		WalletID      string `json:"wallet_id"` // Төлбөр хүлээн авах мерчантын хэтэвч
-		ExpiresMinute int    `json:"expires_minute"`
-		Description   string `json:"description"`
-		Amount        int    `json:"amount"`
-		TransactionID string `json:"transaction_id"` // third party id
-		CallbackURL   string `json:"callback_url"`
-		MetaData      string `json:"meta_data"`
+		WalletID      string  `json:"wallet_id"` // Төлбөр хүлээн авах мерчантын хэтэвч
+		ExpiresMinute int     `json:"expires_minute"`
+		Description   string  `json:"description"`
+		Amount        float64 `json:"amount"`
+		TransactionID string  `json:"transaction_id"` // third party id
+		CallbackURL   string  `json:"callback_url"`
+		MetaData      string  `json:"meta_data"`
 	}
 	InvoiceResponse struct {
 		BaseResponse
@@ -68,7 +68,7 @@ type (
 	InvoiceCheckData struct {
 		InvoiceID      string    `json:"invoice_id"`
 		Status         string    `json:"status"`
-		Amount         int       `json:"amount"`
+		Amount         float64   `json:"amount"`
 		CouponAmount   float64   `json:"coupon_amount"`
 		OriginalAmount float64   `json:"original_amount"`
 		PaidAt         time.Time `json:"paid_at"`
@@ -80,7 +80,7 @@ type (
 	InvoiceCallbackData struct {
 		InvoiceID      string  `json:"invoice_id"`
 		Status         string  `json:"status"`
-		Amount         int     `json:"amount"`
+		Amount         float64 `json:"amount"`
 		CouponAmount   float64 `json:"coupon_amount"`
 		OriginalAmount float64 `json:"original_amount"`
 	}
